requirements.txt
aiohttp
requests
beautifulsoup4
lxml
scholarly
openai
anthropic
numpy
pandas
python-dotenv
chromadb
sentence-transformers